+++ libs/qCC_db/ccPointCloud.h
@@ -708,6 +708,9 @@ public: //other methods
 	//! Exports the specified coordinate dimension(s) to scalar field(s)
 	bool exportCoordToSF(bool exportDims[3]);
 
+	//! Exports the specified normal dimension(s) to scalar field(s)
+	bool exportNormalToSF(bool exportDims[3]);
+
 protected:
 
 	//inherited from ccHObject